Took ben's suggestion to go for longs!

+++ Exercises/04_Week_4/01_Fibonacci/Fibonacci_01/Fibonacci_01.pde
@@ -1,18 +1,20 @@
+// took ben's suggestion to use longs, that way there's no overflow.
+
 void setup() {
   println(fibonacci(100));
 
 }
 
-float fibonacci(int n) {
+long fibonacci(long n) {
 
-  float prevValue = 1;
-  float curValue = 1;
+  long prevValue = 1;
+  long curValue = 1;
 
   if (n <= 1) {
     return n;
   }
   for (int i = 2; i<n; i++) {
-    float temp = curValue;
+    long temp = curValue;
     curValue += prevValue;
     prevValue = temp;
   }
